+++ jax/_src/nn/functions.py
@@ -663,7 +663,7 @@ def _one_hot(x: Array, num_classes: int, *,
   try:
     output_pos_axis = util.canonicalize_axis(axis, x.ndim + 1)
   except TypeError:
-    axis_size = lax.psum(1, axis)
+    axis_size = lax.axis_size(axis)
     if num_classes != axis_size:
       raise ValueError(f"Expected num_classes to match the size of axis {axis}, "
                        f"but {num_classes} != {axis_size}") from None