+++ tests/export_test.py
@@ -1343,7 +1343,7 @@ class JaxExportTest(jtu.JaxTestCase):
         shard_map, mesh=mesh,
         in_specs=(P("x", None),), out_specs=P("x", None))
     def f_jax(b):  # b: f32[2, 4]
-      axis_size = lax.psum(1, "x")
+      axis_size = lax.axis_size("x")
       perm = [(j, (j + 1) % axis_size) for j in range(axis_size)]
       return lax.ppermute(b, "x", perm=perm)